# PERF: get_loc